HomeWork_14/pyproject.toml
[tool.poetry]
name = "homework-13"
version = "0.1.0"
description = ""
authors = ["YpL68 <107142723+YpL68@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "homework_12"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
sqlalchemy = "^2.0.9"
psycopg2 = "^2.9.6"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
alembic = "^1.10.3"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
python-dotenv = "^1.0.0"
fastapi-mail = "^1.2.8"
redis = "^4.5.4"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.32.0"
httpx = "^0.24.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
sphinx = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
